refactor: move shared types in `ic-btc-types` crate (#244)
There are a few types that will be shared across different crates
(`test-utils`, `canister`, `validation`, etc.)

To facilitate the sharing of these types, this commit moves them into a
separate crate. More refactoring will be done in future commits to use
these types across these crates.
