+++ canister/Cargo.toml
@@ -11,6 +11,7 @@ candid = { workspace = true }
 ciborium = { workspace = true }
 hex = { workspace = true }
 ic-btc-interface = { workspace = true }
+ic-btc-types = { workspace = true }
 ic-btc-validation = { workspace = true }
 ic-cdk = { workspace = true }
 ic-cdk-macros = { workspace = true }
@@ -29,6 +30,7 @@ async-std = { version = "1.12.0", features = ["attributes"] }
 bitcoin = { workspace = true, features = ["rand"] } # needed for generating secp256k1 keys.
 byteorder = { workspace = true }
 ic-btc-test-utils = { workspace = true }
+ic-btc-types = { workspace = true, features = ["mock_difficulty"] }
 maplit = "1.0.2"
 proptest = "0.9.4"
 tempfile = "3.2.0"
