+++ canister/src/api/get_balance.rs
@@ -103,10 +103,10 @@ mod test {
     use crate::{
         genesis_block, state,
         test_utils::{random_p2pkh_address, BlockBuilder, TransactionBuilder},
-        types::OutPoint,
         with_state_mut,
     };
     use ic_btc_interface::{Config, Fees, Network};
+    use ic_btc_types::OutPoint;
 
     #[test]
     fn error_on_malformed_address() {
