+++ canister/src/api/get_utxos.rs
@@ -2,10 +2,11 @@ use crate::{
     blocktree::BlockChain,
     charge_cycles,
     runtime::{performance_counter, print},
-    types::{Address, Block, BlockHash, GetUtxosRequest, OutPoint, Page, Txid, Utxo},
+    types::{Address, GetUtxosRequest, Page, Utxo},
     unstable_blocks, verify_has_enough_cycles, with_state, with_state_mut, State,
 };
 use ic_btc_interface::{GetUtxosError, GetUtxosResponse, Utxo as PublicUtxo, UtxosFilter};
+use ic_btc_types::{Block, BlockHash, OutPoint, Txid};
 use serde_bytes::ByteBuf;
 use std::str::FromStr;
 
@@ -275,11 +276,11 @@ mod test {
             random_p2pkh_address, random_p2tr_address, random_p2wpkh_address, random_p2wsh_address,
             BlockBuilder, BlockChainBuilder, TransactionBuilder,
         },
-        types::Block,
         with_state_mut,
     };
     use ic_btc_interface::{Config, Fees, Network};
     use ic_btc_interface::{OutPoint, Utxo};
+    use ic_btc_types::Block;
     use proptest::prelude::*;
 
     #[test]
@@ -551,7 +552,7 @@ mod test {
             .with_transaction(coinbase_tx.clone())
             .build();
         let tx = TransactionBuilder::new()
-            .with_input(crate::types::OutPoint::new(coinbase_tx.txid(), 0))
+            .with_input(ic_btc_types::OutPoint::new(coinbase_tx.txid(), 0))
             .with_output(&address_2, 1000)
             .build();
         let block_1 = BlockBuilder::with_prev_header(block_0.header())
@@ -733,7 +734,7 @@ mod test {
 
         // Extend block 0 with block 1 that spends the 1000 satoshis and gives them to address 2.
         let tx = TransactionBuilder::new()
-            .with_input(crate::types::OutPoint::new(coinbase_tx.txid(), 0))
+            .with_input(ic_btc_types::OutPoint::new(coinbase_tx.txid(), 0))
             .with_output(&address_2, 1000)
             .build();
         let block_1 = BlockBuilder::with_prev_header(block_0.header())
@@ -783,7 +784,7 @@ mod test {
         // Extend block 0 (again) with block 1 that spends the 1000 satoshis to address 3
         // This causes a fork.
         let tx = TransactionBuilder::new()
-            .with_input(crate::types::OutPoint::new(coinbase_tx.txid(), 0))
+            .with_input(ic_btc_types::OutPoint::new(coinbase_tx.txid(), 0))
             .with_output(&address_3, 1000)
             .build();
         let block_1_prime = BlockBuilder::with_prev_header(block_0.header())
@@ -835,7 +836,7 @@ mod test {
         // In this case, the fork of [block 1', block 2'] will be considered the "main"
         // chain, and will be part of the UTXOs.
         let tx = TransactionBuilder::new()
-            .with_input(crate::types::OutPoint::new(tx.txid(), 0))
+            .with_input(ic_btc_types::OutPoint::new(tx.txid(), 0))
             .with_output(&address_4, 1000)
             .build();
         let block_2_prime = BlockBuilder::with_prev_header(block_1_prime.header())
@@ -995,7 +996,7 @@ mod test {
             .with_transaction(coinbase_tx.clone())
             .build();
         let tx = TransactionBuilder::new()
-            .with_input(crate::types::OutPoint::new(coinbase_tx.txid(), 0))
+            .with_input(ic_btc_types::OutPoint::new(coinbase_tx.txid(), 0))
             .with_output(&address_2, 1000)
             .build();
         let block_1 = BlockBuilder::with_prev_header(block_0.header())
