+++ canister/src/unstable_blocks/next_block_headers.rs
@@ -1,6 +1,6 @@
-use crate::types::BlockHash;
 use bitcoin::BlockHeader;
 use ic_btc_interface::Height;
+use ic_btc_types::BlockHash;
 use serde::{Deserialize, Serialize};
 use std::collections::BTreeMap;
 
@@ -64,10 +64,8 @@ impl NextBlockHeaders {
 
 #[cfg(test)]
 mod test {
-    use crate::{
-        test_utils::BlockBuilder, types::BlockHash,
-        unstable_blocks::next_block_headers::NextBlockHeaders,
-    };
+    use crate::{test_utils::BlockBuilder, unstable_blocks::next_block_headers::NextBlockHeaders};
+    use ic_btc_types::BlockHash;
 
     #[test]
     fn test_get_max_height() {
