--- test-utils/BUILD.bazel
@@ -1,18 +1,0 @@
-load("@rules_rust//rust:defs.bzl", "rust_library", "rust_test")
-
-package(default_visibility = ["//visibility:public"])
-
-rust_library(
-    name = "test-utils",
-    srcs = glob(["src/**"]),
-    crate_name = "ic_btc_test_utils",
-    version = "0.1.0",
-    deps = [
-        "@crate_index//:bitcoin",
-    ],
-)
-
-rust_test(
-    name = "ic_btc_test_utils_test",
-    crate = ":test-utils",
-)
